Create step component for welcome page
